[blockchain]添加了货币单位的转换类，默认返回单位为ETH防止余额位数过大

+++ drm-blockchain/src/main/kotlin/service/AccountService.kt
@@ -2,6 +2,7 @@ package moe._47saikyo.service
 
 import org.web3j.crypto.WalletFile
 import org.web3j.tx.TransactionManager
+import java.math.BigDecimal
 import java.math.BigInteger
 
 
@@ -13,7 +14,7 @@ import java.math.BigInteger
 interface AccountService {
     fun newAccount(password: String):WalletFile
     fun newAccountByPersonal(password: String): String
-    fun getBalance(addr: String): BigInteger
+    fun getBalance(addr: String): BigDecimal
     fun chargeFromBank(addr: String, value: String): Boolean
     fun getTxManager(pwd: String, keyFileContent: String): TransactionManager
 }
